+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Removed unused `:resource` param in `Assent.Strategy.AzureAD`
 * Added "email profile" to scope in `Assent.Strategy.AzureAD`
 * Use `response_mode=form_post` for `Assent.Strategy.AzureAD`
+* Updated `Assent.Strategy.OAuth2` to handle access token request correctly when `:auth_method` is `nil` per RFC specs
 
 ## v0.1.4 (2019-11-09)